pyproject.toml
[project]
name = "mcp-zernio"
version = "0.1.0"
description = "FastMCP wrapper for Zernio social media management API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=3.2.0",
    "httpx>=0.28.1",
    "pydantic-settings>=2.13.1",
]

[project.scripts]
mcp-zernio = "zernio_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/zernio_mcp"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "respx>=0.22.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
